requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
app==0.0.1
appnope==0.1.3
approx==0.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
audioread==3.0.1
awscli==1.29.14
Babel==2.12.1
backcall==0.2.0
banal==1.0.6
base==0.0.0
bases==0.2.1
bce-python-sdk==0.8.87
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.28.14
botocore==1.31.14
bottle==0.12.25
boxes==0.0.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
ci-info==0.3.0
click==8.1.6
cloudpickle==2.2.1
cmake==3.27.0
colorama==0.4.4
colorlog==6.7.0
comm==0.1.3
common==0.1.2
configobj==5.0.8
configparser==6.0.1
contextlib2==21.6.0
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cycler==0.10.0
Cython==3.0.0
cython-bbox==0.1.3
darknet==0.3
data==0.4
dataset==1.6.2
dataset-utils==0.0.48
datasets==2.14.1
debugpy==1.6.7
decorator==5.1.1
deep-sort-realtime==1.3.2
defusedxml==0.7.1
dill==0.3.4
dlib==19.24.2
dm-tree==0.1.8
docker==3.4.0
docker-pycreds==0.4.0
docutils==0.16
dropbox==11.36.2
dual==0.1.0
dynamo3==0.4.10
easydict==1.10
easyocr==1.7.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.100.1
fastjsonschema==2.18.0
fastreid==1.4.0
filelock==3.12.2
filterpy==1.4.5
fire==0.5.0
fitz==0.0.1.dev2
Flask==2.3.2
Flask-Babel==2.0.0
flatbuffers==23.5.26
flywheel==0.5.4
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
funcsigs==1.0.2
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.10.0
helpdev==0.7.1
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.23.2
idna==2.10
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
imutils==0.5.4
ipdb==0.13.13
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jax==0.4.14
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
JSON-minify==0.3.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kapre==0.3.7
keras==3.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==8.7
keyrings.alt==5.0.0
kiwisolver==1.4.5
lazy_loader==0.3
Levenshtein==0.21.1
libclang==16.0.6
librosa==0.10.1
llvmlite==0.40.1
lmdb==1.4.1
loguru==0.7.0
looseversion==1.3.0
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.3.2
more-itertools==10.0.0
motmetrics==1.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.12.2
namex==0.0.7
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.7
networkx==3.1
nibabel==5.2.1
ninja==1.11.1
nipype==1.8.6
notebook==7.0.0
notebook_shim==0.2.3
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
object-detection==0.0.3
onnx==1.15.0
opencv-contrib-python==4.6.0.66
opencv-python==4.10.0.84
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
openvino==2023.0.1
opt-einsum==3.3.0
optree==0.10.0
overrides==7.3.1
packaging==23.1
paddle-bfloat==0.1.7
paddle2onnx==1.0.5
paddlefsl==1.1.0
paddlenlp==2.5.2
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pdf2docx==0.5.6
peewee==3.16.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
ply==3.11
pooch==1.8.1
premailer==3.10.0
progressbar==2.5
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
prov==2.0.0
prox==0.0.17
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
pydot==2.0.0
Pygments==2.15.1
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
PyQt5-stubs==5.15.6.0
pyrsistent==0.19.3
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-geoip-python3==1.3
python-json-logger==2.0.7
python-Levenshtein==0.21.1
python-multipart==0.0.6
python-utils==3.7.0
pytz==2024.1
PyWavelets==1.4.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.0
QDarkStyle==3.1
qtconsole==5.4.3
QtPy==2.3.1
rapidfuzz==3.1.2
rarfile==4.0
rdflib==7.0.0
recordclass==0.19.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
roboflow==1.1.2
rpds-py==0.9.2
rsa==4.7.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sahi==0.11.14
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.12.0
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.1.99
seqeval==1.2.2
shapely==2.0.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
SQLAlchemy==1.4.49
stack-data==0.6.2
starlette==0.27.0
stone==3.3.1
supervision==0.19.0
sympy==1.12
tabulate==0.9.0
tci==0.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-addons==0.21.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.14.0
tensorflow-probability==0.20.1
termcolor==2.3.0
terminado==0.17.1
terminaltables==3.1.10
tesseract==0.1.3
tf2-yolov4==0.1.0
Theano==1.0.5
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.18
tight==0.1.0
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.3.2
tsheets @ git+https://github.com/TSheetsTeam/api_python.git@301666970a2b70bb43b712923b8e48c412fda13a
typeguard==2.13.3
typer==0.9.0
typing-validation==1.0.0.post2
typing_extensions==4.10.0
tzdata==2023.3
ultralytics==8.2.32
ultralytics-thop==0.2.8
uri-template==1.3.0
urllib3==1.26.16
utils==1.0.1
uvicorn==0.23.1
visualdl==2.4.2
voluptuous==0.13.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==4.0.8
wrapt==1.14.1
xmltodict==0.13.0
xxhash==3.3.0
yacs==0.1.8
yarl==1.9.2
yolo==0.3.2
yolov4==3.2.0
yolov5==7.0.12
yolov7==0.0.1
yolov7detect==1.0.1
zipp==3.16.2
